BackendText/requirements.txt
flask
flask-cors
python-dotenv
boto3
pillow
torch
transformers
diffusers